scoot non-ajax attractions calls from firebase into attractions.js, fixed resulting problems
